✨ Adds JWT based GitHub App authorization
GitHub App installation token generation is based on JWTs that are
generated with the private key of the GitHub App.

I implemented basic functionality to load a DER RSA private key from a
PEM encoded private key.
We use the jsonwebtoken crate to generate a JWT based with that key.

With that JWT we can now successfully authn and authz.
To verify the process we implemented `get_app` which requires a JWT
based authentication.